HttpRequestIntent as interface #1053
