+++ commercetools-sdk-base/src/main/java/io/sphere/sdk/client/HttpRequestIntent.java
@@ -1,7 +1,6 @@
 package io.sphere.sdk.client;
 
 import io.sphere.sdk.http.*;
-import io.sphere.sdk.models.Base;
 
 import javax.annotation.Nullable;
 import java.io.File;
@@ -11,62 +10,35 @@ import static io.sphere.sdk.http.HttpHeaders.CONTENT_TYPE;
 /**
  * Expresses the http request domain model form commercetools platform as a draft for {@link HttpRequest}.
  */
-public final class HttpRequestIntent extends Base {
-    private final HttpMethod httpMethod;
-    private final String path;
-    private final HttpHeaders headers;
-    @Nullable
-    private final HttpRequestBody body;
+public interface HttpRequestIntent {
+    HttpHeaders getHeaders();
 
-    private HttpRequestIntent(final HttpMethod httpMethod, final String path, final HttpHeaders headers, @Nullable final HttpRequestBody body) {
-        this.headers = headers;
-        this.httpMethod = httpMethod;
-        this.path = path;
-        this.body = body;
-    }
+    HttpMethod getHttpMethod();
 
-    public HttpHeaders getHeaders() {
-        return headers;
-    }
-
-    public HttpMethod getHttpMethod() {
-        return httpMethod;
-    }
-
-    public String getPath() {
-        return path;
-    }
+    String getPath();
 
     @Nullable
-    public HttpRequestBody getBody() {
-        return body;
-    }
+    HttpRequestBody getBody();
 
-    public HttpRequestIntent plusHeader(final String name, final String value) {
-        return HttpRequestIntent.of(getHttpMethod(), getPath(), getHeaders().plus(name, value), getBody());
-    }
+    HttpRequestIntent plusHeader(final String name, final String value);
 
-    public HttpRequestIntent prefixPath(final String prefix) {
-        return HttpRequestIntent.of(getHttpMethod(), prefix + getPath(), getHeaders(), getBody());
-    }
+    HttpRequestIntent prefixPath(final String prefix);
 
-    public HttpRequest toHttpRequest(final String baseUrl) {
-        return HttpRequest.of(getHttpMethod(), baseUrl + getPath(), getHeaders(), getBody());
-    }
+    HttpRequest toHttpRequest(final String baseUrl);
 
-    public static HttpRequestIntent of(final HttpMethod httpMethod, final String path) {
+    static HttpRequestIntent of(final HttpMethod httpMethod, final String path) {
         return of(httpMethod, path, HttpHeaders.of(), null);
     }
 
-    public static HttpRequestIntent of(final HttpMethod httpMethod, final String path, final HttpHeaders headers, @Nullable final HttpRequestBody body) {
-        return new HttpRequestIntent(httpMethod, path, headers, body);
+    static HttpRequestIntent of(final HttpMethod httpMethod, final String path, final HttpHeaders headers, @Nullable final HttpRequestBody body) {
+        return new HttpRequestIntentImpl(httpMethod, path, headers, body);
     }
 
-    public static HttpRequestIntent of(final HttpMethod httpMethod, final String path, final String body) {
+    static HttpRequestIntent of(final HttpMethod httpMethod, final String path, final String body) {
         return of(httpMethod, path, HttpHeaders.of(), StringHttpRequestBody.of(body));
     }
 
-    public static HttpRequestIntent of(final HttpMethod httpMethod, final String path, final File body, final String contentType) {
+    static HttpRequestIntent of(final HttpMethod httpMethod, final String path, final File body, final String contentType) {
         return of(httpMethod, path, HttpHeaders.of(CONTENT_TYPE, contentType), FileHttpRequestBody.of(body));
     }
 }
